+++ terraform/cloudwatch/variables.tf
@@ -1,0 +1,5 @@
+variable "region" {
+  description = "The region to deploy the RDS instances in"
+  default     = "eu-central-1"
+}
+variable "profile" {}